gradio app files/requirements.txt

numpy==1.26.4
Pillow==10.3.0
tensorflow==2.16.1

